chore(ui): trim dead code from models page rework
Remove unused traits, page-layout picker branches, and the standalone cleanup-picker file; consolidate Freestyle constants and shared modal chrome.
