+++ apps/electron/src/renderer/src/locales/en.json
@@ -561,14 +561,11 @@
       "transcription": "Transcription",
       "cleanup": "AI cleanup",
       "freestyleTranscribe": "Freestyle Transcribe",
-      "freestyleCleanup": "Freestyle Cleanup",
       "recommended": "Recommended",
       "managedSignedIn": "Managed cloud · signed in",
       "managedSignIn": "Managed cloud · sign in once, no API key",
       "freestyleBundleSignedIn": "Managed cloud · transcription and AI cleanup included",
       "freestyleBundleSignIn": "Managed cloud · sign in once · transcription and cleanup included",
-      "cleanupDescSignedIn": "Managed cloud · fixes punctuation and filler words",
-      "cleanupDescSignIn": "Managed cloud · sign in once, no API key",
       "onDevice": "On {{phrase}}",
       "yourApiKey": "Your API key",
       "unavailableOnDevice": "Unavailable on this device",
@@ -576,7 +573,6 @@
       "cloudModelCount": "{{count}} cloud models",
       "byokProviders": "Groq, OpenAI, and more",
       "ollamaHint": "Ollama, LM Studio…",
-      "modelWarming": "Model warming",
       "browseLocalVoice": "Browse on-device transcription models",
       "browseByokVoice": "Browse cloud transcription models with your API key",
       "browseLocalCleanup": "Browse on-device cleanup models",
