+++ apps/electron/src/renderer/src/pages/models/transcription-picker.tsx
@@ -2,11 +2,13 @@ import { Button } from "@renderer/components/ui/button";
 import { useCloudAuth } from "@renderer/lib/auth-context";
 import type { AvailableModel } from "@renderer/lib/models";
 import { cn, ON_DEVICE_PHRASE } from "@renderer/lib/utils";
-import { Check, Cloud, ExternalLink, Key, Laptop, Mic, X } from "lucide-react";
+import { Check, Cloud, ExternalLink, Key, Laptop, Mic } from "lucide-react";
 import { useTranslation } from "react-i18next";
-
-import { Eyebrow } from "./page-chrome";
-import { PickerOption } from "./picker-option";
+import {
+  PICKER_MODAL_BODY,
+  PickerModalHeader,
+  PickerOption,
+} from "./picker-option";
 import type { ConfiguredModel } from "./types";
 import type { UseModels } from "./use-models";
 import { displayName } from "./utils";
@@ -19,15 +21,19 @@ export const FREESTYLE_CLOUD_TIER: AvailableModel = {
   type: "voice",
 };
 
+export const FREESTYLE_CLOUD_CLEANUP: AvailableModel = {
+  provider_id: "freestyle-cloud",
+  provider_name: "Freestyle Transcribe",
+  model_id: "freestyle-cloud/post-process",
+  model_name: "Freestyle Cleanup",
+  type: "llm",
+};
+
 const MANAGED_PROVIDERS = new Set([
   FREESTYLE_CLOUD_TIER.provider_id,
   "local-whisper",
   "local-mlx",
 ]);
-
-const PICKER_MODAL_BODY = "space-y-5 px-6 py-6";
-const PICKER_MODAL_HEADER =
-  "border-border flex shrink-0 items-center gap-3 border-b px-6 py-4";
 
 export function recommendedVoiceKey(
   items: { key: string; localEngine?: string }[],
@@ -54,26 +60,20 @@ function voiceMatches(
   return voice?.provider === provider && voice?.model_id === modelId;
 }
 
-/**
- * Minimal transcription picker: Freestyle → browse local → browse BYOK.
- */
+/** Modal tier picker: Freestyle bundle → browse local → browse BYOK. */
 export function TranscriptionPicker({
   m,
-  layout,
   onClose,
   onPickCloud,
   onBrowseLocal,
   onBrowseCloud,
-  onConfigureWarming,
   busy,
 }: {
   m: UseModels;
-  layout: "page" | "modal";
-  onClose?: () => void;
+  onClose: () => void;
   onPickCloud: (model: AvailableModel) => void;
   onBrowseLocal: () => void;
   onBrowseCloud: () => void;
-  onConfigureWarming?: () => void;
   busy?: boolean;
 }): React.JSX.Element {
   const { t } = useTranslation();
@@ -107,113 +107,73 @@ export function TranscriptionPicker({
       ? t("models.picker.cloudModelCount", { count: byokCount })
       : t("models.picker.byokProviders");
 
-  const body = (
-    <div className={cn(layout === "page" ? "space-y-5" : PICKER_MODAL_BODY)}>
-      {layout === "page" && (
-        <div className="flex items-center justify-between gap-3">
-          <Eyebrow
-            text={t("models.picker.transcription")}
-            accent
-            mono={false}
-          />
-          {onConfigureWarming && (
-            <Button
-              variant="link"
-              size="sm"
-              onClick={onConfigureWarming}
-              className="text-muted-foreground h-auto px-0 text-[12px]"
-            >
-              {t("models.picker.modelWarming")}
-            </Button>
+  return (
+    <>
+      <PickerModalHeader
+        icon={Mic}
+        title={t("models.picker.transcription")}
+        onClose={onClose}
+      />
+      <div className={PICKER_MODAL_BODY}>
+        <button
+          type="button"
+          disabled={busy}
+          onClick={() => onPickCloud(FREESTYLE_CLOUD_TIER)}
+          className={cn(
+            "border-border hover:border-primary/35 w-full rounded-[14px] border p-6 text-left transition-[transform,border-color,background-color] duration-150 ease-out active:scale-[0.99] disabled:pointer-events-none disabled:opacity-60",
+            freestyleSelected
+              ? "border-primary/45 bg-primary/[0.06]"
+              : "bg-primary/[0.03]",
           )}
-        </div>
-      )}
-
-      <button
-        type="button"
-        disabled={busy}
-        onClick={() => onPickCloud(FREESTYLE_CLOUD_TIER)}
-        className={cn(
-          "border-border hover:border-primary/35 w-full rounded-[14px] border p-6 text-left transition-[transform,border-color,background-color] duration-150 ease-out active:scale-[0.99] disabled:pointer-events-none disabled:opacity-60",
-          freestyleSelected
-            ? "border-primary/45 bg-primary/[0.06]"
-            : "bg-primary/[0.03]",
-        )}
-      >
-        <div className="flex items-start gap-4">
-          <span className="bg-primary/10 text-primary flex size-10 shrink-0 items-center justify-center rounded-[10px]">
-            <Cloud className="size-4" />
-          </span>
-          <div className="min-w-0 flex-1">
-            <div className="flex flex-wrap items-center gap-2">
-              <span className="text-foreground text-[15px] font-semibold tracking-[-0.01em]">
-                {t("models.picker.freestyleTranscribe")}
-              </span>
-              {!freestyleSelected && (
-                <span className="text-primary text-[10px] font-semibold uppercase tracking-wide">
-                  {t("models.picker.recommended")}
+        >
+          <div className="flex items-start gap-4">
+            <span className="bg-primary/10 text-primary flex size-10 shrink-0 items-center justify-center rounded-[10px]">
+              <Cloud className="size-4" />
+            </span>
+            <div className="min-w-0 flex-1">
+              <div className="flex flex-wrap items-center gap-2">
+                <span className="text-foreground text-[15px] font-semibold tracking-[-0.01em]">
+                  {t("models.picker.freestyleTranscribe")}
                 </span>
-              )}
+                {!freestyleSelected && (
+                  <span className="text-primary text-[10px] font-semibold uppercase tracking-wide">
+                    {t("models.picker.recommended")}
+                  </span>
+                )}
+              </div>
+              <p className="text-muted-foreground mt-1.5 text-[13px] leading-relaxed">
+                {cloud.user
+                  ? t("models.picker.freestyleBundleSignedIn")
+                  : t("models.picker.freestyleBundleSignIn")}
+              </p>
             </div>
-            <p className="text-muted-foreground mt-1.5 text-[13px] leading-relaxed">
-              {cloud.user
-                ? t("models.picker.freestyleBundleSignedIn")
-                : t("models.picker.freestyleBundleSignIn")}
-            </p>
+            {freestyleSelected && (
+              <Check className="text-primary mt-1 size-[18px] shrink-0" />
+            )}
           </div>
-          {freestyleSelected && (
-            <Check className="text-primary mt-1 size-[18px] shrink-0" />
-          )}
+        </button>
+
+        <div className="border-border divide-border overflow-hidden rounded-[12px] border divide-y">
+          <PickerOption
+            icon={Laptop}
+            title={t("models.picker.onDevice", { phrase: ON_DEVICE_PHRASE })}
+            hint={localHint}
+            active={localActive}
+            onClick={onBrowseLocal}
+            browseLabel={t("models.picker.browseLocalVoice")}
+          />
+          <PickerOption
+            icon={Key}
+            title={t("models.picker.yourApiKey")}
+            hint={byokLabel}
+            active={byokActive}
+            onClick={onBrowseCloud}
+            browseLabel={t("models.picker.browseByokVoice")}
+          />
         </div>
-      </button>
-
-      <div className="border-border divide-border overflow-hidden rounded-[12px] border divide-y">
-        <PickerOption
-          icon={Laptop}
-          title={t("models.picker.onDevice", { phrase: ON_DEVICE_PHRASE })}
-          hint={localHint}
-          active={localActive}
-          onClick={onBrowseLocal}
-          browseLabel={t("models.picker.browseLocalVoice")}
-        />
-        <PickerOption
-          icon={Key}
-          title={t("models.picker.yourApiKey")}
-          hint={byokLabel}
-          active={byokActive}
-          onClick={onBrowseCloud}
-          browseLabel={t("models.picker.browseByokVoice")}
-        />
       </div>
-    </div>
+    </>
   );
-
-  if (layout === "modal") {
-    return (
-      <>
-        <header className={PICKER_MODAL_HEADER}>
-          <Mic className="text-muted-foreground h-3.5 w-3.5 shrink-0" />
-          <span className="text-foreground flex-1 text-[13px] font-semibold">
-            {t("models.picker.transcription")}
-          </span>
-          {onClose && (
-            <Button
-              variant="ghost"
-              size="icon-sm"
-              onClick={onClose}
-              className="shrink-0"
-              aria-label="Close"
-            >
-              <X />
-            </Button>
-          )}
-        </header>
-        {body}
-      </>
-    );
-  }
-
-  return body;
 }
 
 export function OpenModelSourceButton({
